Handle corrupt or empty database records (#78)
* Handle corrupt or empty database records

In cases where a crash record file is corrupted or empty,
BacktraceDatabaseRecord.readFromFile will return null

Adds handling for this case by treating it as an invalid record

* Add unit test, fix checking null record

+++ backtrace-library/src/main/java/backtraceio/library/BacktraceDatabase.java
@@ -415,6 +415,10 @@ public class BacktraceDatabase implements Database {
 
         for (File file : files) {
             BacktraceDatabaseRecord record = BacktraceDatabaseRecord.readFromFile(file);
+            if (record == null) {
+                continue;
+            }
+
             if (!record.valid()) {
                 record.delete();
                 continue;
